backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Werkzeug
SQLAlchemy
python-dotenv
psycopg2-binary